+++ src/cslf_int.js
@@ -1,7 +1,7 @@
-import './styles/psb_int_styles.css';
 import './styles/spinner.css';
+import './styles/styles.css';
 
-const template = require('./js/psb_int4.hbs')
+const template = require('./js/cslf_int.hbs')
 
 const fadeOutEffect = () => {
   var fadeTarget = document.getElementById("overlay");