+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Change Log
+## September 6, 2022
+- Rebuilt Webpack config to allow for multi-page app, merged branches
 ## September 4, 2022
 - Full-draft of 4-term version ready to test on pei-test1
 ## August 10, 2022